pyproject.toml
[tool.poetry]
name = "atvd-02"
version = "0.1.0"
description = "Atividate 02 - Extenção I"
authors = ["Wellington Casarin"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
python-docx = "^1.1.2"
odfpy = "^1.4.1"
pypandoc-binary = "^1.14"
jinja2 = "^3.1.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
